Added ball hit sound
Balls now make a sound on collision with other objects. Its volume varies with the relative velocity of the collision and the distance to the listener (camera).

I tried implementing sounds for ice and rock collisions but ran into trouble. I believe the problem has to do with playing too many sounds at once.