Initial work on ball defense level
Ball defense level! Ball catcher holds the ball in the air and the boulder spawner tries to knock it away. Very rough so far.

Also the player movement is changed a bit. The player has lower drag in the physics engine but some manually applied drag in the movement script.